+++ client/src/Regula.OpenApi.WebClient/Model/TextFieldValue.cs
@@ -111,7 +111,7 @@ namespace Regula.OpenApi.WebClient.Model
         /// </summary>
         /// <value>Page index of the image from input list</value>
         [DataMember(Name="pageIndex", EmitDefaultValue=true)]
-        public int PageIndex { get; set; }
+        public long PageIndex { get; set; }
 
         /// <summary>
         /// Min recognition probability. Combined minimum probability from single characters probabilities
